fix "pod spec lint" : define menuView and few other class members as private property
menuView, oldAngle, currentAngle and isPresentedFirst
are used inside the collision.action block. Instead we can
define them in the .m file so they remain as private
property but we can access them using strongSelf

this is done to resolve the warnings caused by using self in the collision
action block.
warning: capturing 'self' strongly in this block is likely to lead to a retain cycle [-Warc-retain-cycles]